Return & use error code from OpenVR initialisation
This is part of https://github.com/webaverse/metachromium-bin/issues/12.

Previously, when the OpenVR initialisation failed, we didn't do anything
with the error code.

This commit changes the behaviour of the qrvr C++ module to return the
error code received from OpenVR, and the JS will exit the Node process
with the returned error code, if non-0.